refactor: avgrating int to double

+++ gallerygather/gallerygather/src/main/java/com/kosa/gallerygather/dto/ExhibitionDto.java
@@ -26,7 +26,7 @@ public class ExhibitionDto {
     private int likeCount;
     private int readCount;
     private int reviewCount;
-    private long avgRating;
+    private double avgRating;
 
     public ExhibitionDto (Exhibition exhibition) {
         this.id = exhibition.getId();
